refactor manifold
