+++ rubato/utils/defaults.py
@@ -159,9 +159,9 @@ class Defaults:
         # Whether this hitbox is just a trigger or not.
         "scale": 1,  #. . . . . . . . . . . . . . . . . . . . . . . int
         # The scale of the polygon
-        "on_collide": lambda colInfo: None,  #. . . . . . . . . . . Callable
+        "on_collide": lambda Manifold: None,  #. . . . . . . . . . . Callable
         # The on_collide function to call when a collision happens with this hitbox
-        # Must take a rb.ColInfo.
+        # Must take a rb.Manifold.
         "color": None,  # . . . . . . . . . . . . . . . . . . . . . Color
         # The color to fill this hitbox with.
         "tag": "",  # . . . . . . . . . . . . . . . . . . . . . . . str
